requirements.txt
Django==1.10.5
djangorestframework==3.5.3
olefile==0.44
Pillow==2.6.1
wheel==0.24.0
